// src/app/models/passwordChangeRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PasswordChangeRequest { 
    oldPassword?: string;
    newPassword?: string;
}